requirements.txt
Django==3.2.7
psycopg2==2.9.1
django-better-admin-arrayfield
djangorestframework
jsonschema
celery
redis